+++ sugar/lib/src/core/equality.dart
@@ -1,6 +1,5 @@
 import 'package:meta/meta.dart';
 import 'package:sugar/collection.dart';
-import 'package:sugar/core.dart';
 
 /// A skeleton to simplify implementation of symmetrical equality.
 mixin Equality {
@@ -9,12 +8,10 @@ mixin Equality {
   bool operator == (dynamic other) => identical(this, other) || (runtimeType == other.runtimeType && fields.equals(other.fields));
 
   @override
-  int get hashCode => hash(fields);
-
+  int get hashCode => Object.hashAll(fields);
 
   @override
   String toString() => '$runtimeType(${fields.join(', ')})';
-
 
   /// The fields used to determine equality.
   @protected List<dynamic> get fields;