+++ sugar/test/src/collection/equality_test.dart
@@ -1,7 +1,8 @@
 import 'dart:collection';
 import 'dart:typed_data';
-import 'package:test/test.dart';
+
 import 'package:sugar/sugar.dart';
+import 'package:test/test.dart';
 
 void main() {
   const list = [1, 2, 3];