+++ sugar/test/src/collection/iterables_test.dart
@@ -1,5 +1,5 @@
-import 'package:test/test.dart';
 import 'package:sugar/sugar.dart';
+import 'package:test/test.dart';
 
 void main() {
   const list = [1, 2, 3];